Fix storage.js writable store creation
The start function passed to a writable store must return a callable stop function that is called after the last subscriber unsubscribes. It can also return nothing which counts as noop. Here, the start function was async, so it returned a promise. Svelte then tried to call that promise as a function which caused an error.